code/requirements.txt
wget
pandas
#csv
json
psycopg2-binary
sqlalchemy
pymongo
geopy
numpy
prettytable
uszipcode